+++ examples/lazy_and_delay.rb
@@ -5,11 +5,12 @@ $: << File.expand_path('../../lib', __FILE__)
 require 'benchmark'
 
 require 'concurrent/delay'
+require 'concurrent/lazy_reference'
 
 n = 500_000
 
 delay = Concurrent::Delay.new{ nil }
-lazy = Concurrent::Lazy.new{ nil }
+lazy = Concurrent::LazyReference.new{ nil }
 
 delay.value
 lazy.value